refactor(controllers): 로또 결과를 비교하는 서브 컨트롤러 생성
